+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Breaking changes
 - Non-boolean arguments to boolean operators are no longer allowed, this was previously possible due to bug (#1808)
 
+## Deprecated
+ - Leaving a nullable attribute unassigned now produces a deprecation warning. Explicitly assign null instead. (#1775)
+
 ## Fixed
  - Various compiler error reporting improvements (#1810)